Remove support for old NodeJS versions from function serializer
