+++ changelog/pending/20221206--sdk-nodejs--remove-function-serialization-code-for-out-of-suppport-nodejs-versions.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: chore
+  scope: sdk/nodejs
+  description: Remove function serialization code for out of suppport NodeJS versions.
